Move query into models
